refactor: move getTSConfig to util
