+++ src/clean.ts
@@ -16,14 +16,13 @@
 import * as chalk from 'chalk';
 
 import {Options} from './cli';
-import {TSConfig} from './tscfg';
-import {rimrafp} from './util';
+import {getTSConfig, rimrafp} from './util';
 
 /**
  * Remove files generated by the build.
  */
 export async function clean(options: Options): Promise<void> {
-  const tsconfig = (await TSConfig.get(options.targetRootDir)).contents;
+  const tsconfig = (await getTSConfig(options.targetRootDir));
   if (tsconfig.compilerOptions && tsconfig.compilerOptions.outDir) {
     const outDir = tsconfig.compilerOptions.outDir;
     if (outDir === '.') {
